Disconnect PaymentManager on invalid state
PaymentManager can be in an invalid state if methods are called without
first calling Init. This makes PaymentManager disconnect in this case
instead of crashing.

Bug: 367888383
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5967714
Commit-Queue: Nick Navarro <npnavarro@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376412}

+++ content/browser/payments/payment_manager.cc
@@ -18,6 +18,22 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "content/public/browser/service_worker_context.h"
 
 namespace content {
+namespace {
+
+constexpr char kInvalidPaymentManagerStateMessage[] =
+    "Scope URL not set, Init may not have been called.";
+
+enum class ReasonCode : uint32_t {
+  kInvalidContextUrl,
+  kNonUnicodeScopeString,
+  kInvalidScopeUrl,
+  kCrossOriginContextAndScope,
+  kCrossOriginDataAccess,
+  kRenderProcessCannotAccessOrigin,
+  kInvalidState,
+};
+
+}  // namespace
 
 PaymentManager::PaymentManager(
     PaymentAppContextImpl* payment_app_context,
@@ -39,15 +55,6 @@ PaymentManager::~PaymentManager() {
 
 void PaymentManager::Init(const GURL& context_url, const std::string& scope) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
-
-  enum class ReasonCode : uint32_t {
-    kInvalidContextUrl,
-    kNonUnicodeScopeString,
-    kInvalidScopeUrl,
-    kCrossOriginContextAndScope,
-    kCrossOriginDataAccess,
-    kRenderProcessCannotAccessOrigin,
-  };
 
   if (!context_url.is_valid()) {
     receiver_.ResetWithReason(
@@ -91,6 +98,11 @@ void PaymentManager::DeletePaymentInstrument(
     const std::string& instrument_key,
     PaymentManager::DeletePaymentInstrumentCallback callback) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   payment_app_context_->payment_app_database()->DeletePaymentInstrument(
       scope_, instrument_key, std::move(callback));
@@ -100,6 +112,11 @@ void PaymentManager::GetPaymentInstrument(
     const std::string& instrument_key,
     PaymentManager::GetPaymentInstrumentCallback callback) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   payment_app_context_->payment_app_database()->ReadPaymentInstrument(
       scope_, instrument_key, std::move(callback));
@@ -108,6 +125,11 @@ void PaymentManager::GetPaymentInstrument(
 void PaymentManager::KeysOfPaymentInstruments(
     PaymentManager::KeysOfPaymentInstrumentsCallback callback) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   payment_app_context_->payment_app_database()->KeysOfPaymentInstruments(
       scope_, std::move(callback));
@@ -117,6 +139,11 @@ void PaymentManager::HasPaymentInstrument(
     const std::string& instrument_key,
     PaymentManager::HasPaymentInstrumentCallback callback) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   payment_app_context_->payment_app_database()->HasPaymentInstrument(
       scope_, instrument_key, std::move(callback));
@@ -127,6 +154,11 @@ void PaymentManager::SetPaymentInstrument(
     payments::mojom::PaymentInstrumentPtr details,
     PaymentManager::SetPaymentInstrumentCallback callback) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   if (should_set_payment_app_info_) {
     payment_app_context_->payment_app_database()->WritePaymentInstrument(
@@ -144,6 +176,11 @@ void PaymentManager::SetPaymentInstrumentIntermediateCallback(
     PaymentManager::SetPaymentInstrumentCallback callback,
     payments::mojom::PaymentHandlerStatus status) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   if (status != payments::mojom::PaymentHandlerStatus::SUCCESS ||
       !should_set_payment_app_info_) {
@@ -159,12 +196,23 @@ void PaymentManager::SetPaymentInstrumentIntermediateCallback(
 void PaymentManager::ClearPaymentInstruments(
     ClearPaymentInstrumentsCallback callback) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   payment_app_context_->payment_app_database()->ClearPaymentInstruments(
       scope_, std::move(callback));
 }
 
 void PaymentManager::SetUserHint(const std::string& user_hint) {
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
+
   payment_app_context_->payment_app_database()->SetPaymentAppUserHint(
       scope_, user_hint);
 }
@@ -173,6 +221,11 @@ void PaymentManager::EnableDelegations(
     const std::vector<payments::mojom::PaymentDelegation>& delegations,
     PaymentManager::EnableDelegationsCallback callback) {
   DCHECK_CURRENTLY_ON(BrowserThread::UI);
+  if (scope_.is_empty()) {
+    receiver_.ResetWithReason(static_cast<uint32_t>(ReasonCode::kInvalidState),
+                              kInvalidPaymentManagerStateMessage);
+    return;
+  }
 
   payment_app_context_->payment_app_database()->EnablePaymentAppDelegations(
       scope_, delegations, std::move(callback));
